feat(components): Add unsafe message component builders (#7387)
