+++ packages/discord.js/src/index.js
@@ -89,6 +89,7 @@ exports.CommandInteractionOptionResolver = require('./structures/CommandInteract
 exports.ContextMenuCommandInteraction = require('./structures/ContextMenuCommandInteraction');
 exports.DMChannel = require('./structures/DMChannel');
 exports.Embed = require('@discordjs/builders').Embed;
+exports.UnsafeEmbed = require('@discordjs/builders').UnsafeEmbed;
 exports.Emoji = require('./structures/Emoji').Emoji;
 exports.Guild = require('./structures/Guild').Guild;
 exports.GuildAuditLogs = require('./structures/GuildAuditLogs');
@@ -182,8 +183,11 @@ exports.UserFlags = require('discord-api-types/v9').UserFlags;
 exports.WebhookType = require('discord-api-types/v9').WebhookType;
 exports.ActionRow = require('@discordjs/builders').ActionRow;
 exports.ButtonComponent = require('@discordjs/builders').ButtonComponent;
+exports.UnsafeButtonComponent = require('@discordjs/builders').UnsafeButtonComponent;
 exports.SelectMenuComponent = require('@discordjs/builders').SelectMenuComponent;
+exports.UnsafeSelectMenuComponent = require('@discordjs/builders').UnsafeButtonComponent;
 exports.SelectMenuOption = require('@discordjs/builders').SelectMenuOption;
+exports.UnsafeSelectMenuOption = require('@discordjs/builders').UnsafeSelectMenuOption;
 exports.DiscordAPIError = require('@discordjs/rest').DiscordAPIError;
 exports.HTTPError = require('@discordjs/rest').HTTPError;
 exports.RateLimitError = require('@discordjs/rest').RateLimitError;
